web_backend/requirements.txt
fastapi
uvicorn
python-dotenv
stripe
supabase-py